DAPI-1824 Call to community-api for referral send is synchronous and services annotated as transactional.
