+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/DynamicFrameworkContractService.kt
@@ -5,8 +5,10 @@ import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.DynamicFrameworkContract
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.repository.ReferralRepository
 import java.util.UUID
+import javax.transaction.Transactional
 
 @Service
+@Transactional
 class DynamicFrameworkContractService(val referralRepository: ReferralRepository) {
   fun getDynamicFrameworkContractByReferralId(id: UUID): DynamicFrameworkContract? {
     return referralRepository.findByIdOrNull(id)?.intervention?.dynamicFrameworkContract
