+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/NotifyService.kt
@@ -16,6 +16,7 @@ import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events.ReferralEve
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.exception.AsyncEventExceptionHandling
 import java.net.URI
 import java.util.UUID
+import javax.transaction.Transactional
 
 interface NotifyService {
   fun generateResourceUrl(baseURL: String, path: String, id: UUID): URI {
@@ -24,6 +25,7 @@ interface NotifyService {
 }
 
 @Service
+@Transactional
 class NotifyActionPlanService(
   @Value("\${notify.templates.action-plan-submitted}") private val actionPlanSubmittedTemplateID: String,
   @Value("\${interventions-ui.baseurl}") private val interventionsUIBaseURL: String,
@@ -53,6 +55,7 @@ class NotifyActionPlanService(
 }
 
 @Service
+@Transactional
 class NotifyEndOfServiceReportService(
   @Value("\${notify.templates.end-of-service-report-submitted}") private val endOfServiceReportSubmittedTemplateID: String,
   @Value("\${interventions-ui.baseurl}") private val interventionsUIBaseURL: String,
@@ -82,6 +85,7 @@ class NotifyEndOfServiceReportService(
 }
 
 @Service
+@Transactional
 class NotifyAppointmentService(
   @Value("\${notify.templates.appointment-not-attended}") private val appointmentNotAttendedTemplateID: String,
   @Value("\${notify.templates.concerning-behaviour}") private val concerningBehaviourTemplateID: String,
@@ -130,6 +134,7 @@ class NotifyAppointmentService(
 }
 
 @Service
+@Transactional
 class NotifyReferralService(
   @Value("\${notify.templates.referral-sent}") private val referralSentTemplateID: String,
   @Value("\${notify.templates.referral-assigned}") private val referralAssignedTemplateID: String,
